restaurant service impl

+++ identity-service/src/main/java/kz/olzhas/repository/UserCredentialRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserCredentialRepository  extends JpaRepository<UserCredential,Integer> {
+public interface UserCredentialRepository  extends JpaRepository<UserCredential,Long> {
     Optional<UserCredential> findByName(String username);
 }
